LPS-64797 Move
