+++ modules/apps/foundation/petra/petra-mail-util/bnd.bnd
@@ -1,5 +1,5 @@
-Bundle-Name: Liferay Mail Utilities
-Bundle-SymbolicName: com.liferay.mail.util
+Bundle-Name: Liferay Petra Mail Utilities
+Bundle-SymbolicName: com.liferay.petra.mail.util
 Bundle-Version: 1.0.0
 Export-Package: com.liferay.mail.util
 Liferay-Releng-Module-Group-Description:
