+++ modules/apps/foundation/petra/petra-mail-util/build.gradle
