+++ modules/apps/foundation/petra/petra-mail-util/src/main/java/com/liferay/petra/mail/util/JavaMailUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.mail.util;
+package com.liferay.petra.mail.util;
 
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.FileUtil;
