+++ modules/apps/foundation/petra/petra-mail-util/src/main/java/com/liferay/petra/mail/util/LiferayMimeMessage.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.util.mail;
+package com.liferay.petra.mail.util;
 
 import com.liferay.portal.kernel.util.ArrayUtil;
 
@@ -22,7 +22,7 @@ import javax.mail.internet.MimeMessage;
 
 /**
  * @author Jorge Ferrer
- * @see com.liferay.petra.mail.util.LiferayMimeMessage
+ * @see com.liferay.util.mail.LiferayMimeMessage
  */
 public class LiferayMimeMessage extends MimeMessage {
