+++ modules/apps/foundation/petra/petra-mail-util/src/main/java/com/liferay/petra/mail/util/MailEngine.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.mail.util;
+package com.liferay.petra.mail.util;
 
 import com.liferay.mail.kernel.model.Account;
 import com.liferay.mail.kernel.model.FileAttachment;
