+++ modules/apps/foundation/portal/portal-pop-notifications/build.gradle
@@ -6,5 +6,5 @@ dependencies {
 	provided group: "javax.portlet", name: "portlet-api", version: "2.0"
 	provided group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
-	provided project(":apps:foundation:mail:mail-util")
+	provided project(":apps:foundation:petra:petra-mail-util")
 }
