+++ modules/apps/foundation/portal/portal-pop-notifications/src/main/java/com/liferay/portal/pop/notifications/messaging/POPNotificationsMessageListener.java
@@ -15,7 +15,7 @@
 package com.liferay.portal.pop.notifications.messaging;
 
 import com.liferay.mail.kernel.model.Account;
-import com.liferay.mail.util.MailEngine;
+import com.liferay.petra.mail.util.MailEngine;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.messaging.BaseSchedulerEntryMessageListener;
