+++ util-java/src/com/liferay/util/mail/InternetAddressUtil.java
@@ -31,7 +31,7 @@ import org.apache.commons.validator.routines.EmailValidator;
 
 /**
  * @author Alexander Chow
- * @see com.liferay.mail.util.InternetAddressUtil
+ * @see com.liferay.petra.mail.util.InternetAddressUtil
  */
 public class InternetAddressUtil {
