+++ util-java/src/com/liferay/util/mail/MailEngineException.java
@@ -18,7 +18,7 @@ import com.liferay.portal.kernel.exception.NestableException;
 
 /**
  * @author Brian Wing Shun Chan
- * @see com.liferay.mail.util.MailEngineException
+ * @see com.liferay.petra.mail.util.MailEngineException
  */
 public class MailEngineException extends NestableException {
