+++ util-java/src/com/liferay/util/mail/MailServerException.java
@@ -18,7 +18,7 @@ import org.apache.commons.lang.exception.NestableException;
 
 /**
  * @author Alexander Chow
- * @see com.liferay.mail.util.MailServerException
+ * @see com.liferay.petra.mail.util.MailServerException
  */
 public class MailServerException extends NestableException {
